Authorization for editing content

+++ src/main/java/com/santander/bootcamp/comercioacoes/controller/StockController.java
@@ -3,11 +3,14 @@ package com.santander.bootcamp.comercioacoes.controller;
 import com.santander.bootcamp.comercioacoes.data.model.dto.StockDTO;
 import com.santander.bootcamp.comercioacoes.data.service.StockService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
+import javax.validation.constraints.NotEmpty;
+import javax.validation.constraints.NotNull;
 import java.util.List;
 
 @SuppressWarnings("ALL")
@@ -36,18 +39,26 @@ public class StockController {
     }
 
     @PostMapping(value = "/post", consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<StockDTO> postStock(@Valid @RequestBody StockDTO dto) {
-        return ResponseEntity
-                .ok()
-                .contentType( MediaType.APPLICATION_JSON )
-                .body( this.service.save( dto ) );
+    public ResponseEntity<StockDTO> postStock(@Valid @RequestBody StockDTO dto, @NotNull @NotEmpty @RequestParam String authorization) {
+        if ( authorization.compareTo( System.getenv( "EDIT_CONTENT_AUTH_KEY" ) ) != 0 ) {
+            return ResponseEntity.status( HttpStatus.UNAUTHORIZED ).build();
+        } else {
+            return ResponseEntity
+                    .ok()
+                    .contentType( MediaType.APPLICATION_JSON )
+                    .body( this.service.save( dto ) );
+        }
     }
 
     @DeleteMapping(value = "/delete/{id}", produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<StockDTO> deleteStock(@PathVariable Long id) {
-        return ResponseEntity
-                .ok()
-                .contentType( MediaType.APPLICATION_JSON )
-                .body( this.service.deleteByID( id ) );
+    public ResponseEntity<StockDTO> deleteStock(@PathVariable Long id, @NotNull @NotEmpty @RequestParam String authorization) {
+        if ( authorization.compareTo( System.getenv( "EDIT_CONTENT_AUTH_KEY" ) ) != 0 ) {
+            return ResponseEntity.status( HttpStatus.UNAUTHORIZED ).build();
+        } else {
+            return ResponseEntity
+                    .ok()
+                    .contentType( MediaType.APPLICATION_JSON )
+                    .body( this.service.deleteByID( id ) );
+        }
     }
 }
